requirements.txt
selenium>=4.20.0,<5
pytest>=7.4.0,<9
allure-pytest>=2.13.5,<3
